List type params issue fixed
For any endpoint function, whenever there's a parameter of type List<int>, it was directly converted into string and the results were => '[v1,v2,v3....]'.
When the request was sent, the Api returns only one result.
So now, a function is added that joins the list elements like v1,v2,v3... and this works perfectly.